move background store into /background
